// src/modules/simulation/enums/simulation-action.enum.ts
export enum SimulationActionType {
    GO_TO = 'GO_TO',
    CLICK_BY_TEXT = 'CLICK_BY_TEXT',
    WAIT_FOR_SELECTOR = 'WAIT_FOR_SELECTOR',
    WAIT_FOR_NAVIGATION = 'WAIT_FOR_NAVIGATION',
    WAIT_FOR_FUNCTION = 'WAIT_FOR_FUNCTION',
    WAIT_FOR_TIME = 'WAIT_FOR_TIME',
    FILL_INPUT = 'FILL_INPUT',
    SELECT_OPTION = 'SELECT_OPTION',
}